+++ src/AngleSharp/Services/Default/AttributeSelectorFactory.cs
@@ -27,8 +27,9 @@
         /// <param name="name">The name of the attribute.</param>
         /// <param name="value">The value of the attribute.</param>
         /// <param name="prefix">The prefix for the attribute.</param>
+        /// <param name="insensitive">Sets the evaluation mode.</param>
         /// <returns></returns>
-        public delegate ISelector Creator(String name, String value, String prefix);
+        public delegate ISelector Creator(String name, String value, String prefix, Boolean insensitive);
 
         /// <summary>
         /// Registers a new creator for the specified combinator.
@@ -65,8 +66,9 @@
         /// <param name="name">The name of the attribute.</param>
         /// <param name="value">The used value, if any.</param>
         /// <param name="prefix">The given prefix, if any.</param>
+        /// <param name="insensitive">Should be evaluated insensitive.</param>
         /// <returns>The selector with the given options.</returns>
-        protected virtual ISelector CreateDefault(String name, String value, String prefix)
+        protected virtual ISelector CreateDefault(String name, String value, String prefix, Boolean insensitive)
         {
             return SimpleSelector.AttrAvailable(name, value);
         }
@@ -78,17 +80,18 @@
         /// <param name="name">The name of the attribute.</param>
         /// <param name="value">The used value, if any.</param>
         /// <param name="prefix">The given prefix, if any.</param>
+        /// <param name="insensitive">Should be evaluated insensitive.</param>
         /// <returns>The associated selector.</returns>
-        public ISelector Create(String combinator, String name, String value, String prefix)
+        public ISelector Create(String combinator, String name, String value, String prefix, Boolean insensitive)
         {
             var creator = default(Creator);
 
             if (_creators.TryGetValue(combinator, out creator))
             {
-                return creator.Invoke(name, value, prefix);
+                return creator.Invoke(name, value, prefix, insensitive);
             }
 
-            return CreateDefault(name, value, prefix);
+            return CreateDefault(name, value, prefix, insensitive);
         }
     }
 }